backend/requirements.txt
aiosqlite
alembic
anyio
asyncpg
bcrypt
cffi
click
cryptography
#dbus-python
dnspython
ecdsa
email-validator
fastapi
greenlet
#gyp==0.1
h11
httptools
idna
Mako
MarkupSafe
passlib
pyasn1
pycparser
pydantic
#PyGObject
python-dotenv
python-jose
python-multipart
PyYAML
rsa
six
sniffio
SQLAlchemy
starlette
typing_extensions
uvicorn
# uvloop
watchfiles
websockets
psycopg2-binary

tuya-connector-python
tuya-iot-py-sdk
pycryptodomex

# graphviz
# sqlalchemy-visualize
# postgresql-client
